+++ examples/nuxt/nuxt.config.js
@@ -11,10 +11,10 @@ export default {
       {
         hid: 'description',
         name: 'description',
-        content: process.env.npm_package_description || ''
-      }
+        content: process.env.npm_package_description || '',
+      },
     ],
-    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }]
+    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }],
   },
   /*
    ** Customize the progress-bar color
@@ -23,7 +23,7 @@ export default {
   /*
    ** Global CSS
    */
-  css: ['bulmil/dist/css/bulmil.css'],
+  css: ['@bulmil/core/dist/css/bulmil.css'],
   /*
    ** Plugins to load before mounting the App
    */
@@ -38,8 +38,8 @@ export default {
   modules: ['nuxt-stencil'],
 
   stencil: {
-    lib: 'bulmil',
-    prefix: 'bm-'
+    lib: '@bulmil/core',
+    prefix: 'bm-',
   },
   /*
    ** Build configuration
@@ -48,6 +48,6 @@ export default {
     /*
      ** You can extend webpack config here
      */
-    extend(config, ctx) {}
-  }
+    extend(config, ctx) {},
+  },
 }