+++ packages/bulmil-svelte/src/index.ts
@@ -1,7 +1,0 @@
-import { defineCustomElements } from 'bulmil/dist/loader';
-
-export * from './proxies';
-
-if (typeof window !== 'undefined') {
-  defineCustomElements(window);
-}