+++ packages/core/src/core/layout/footer/test/footer.spec.tsx
@@ -1,0 +1,19 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { Footer } from '../footer';
+
+it('Should render properly', async (): Promise<void> => {
+  const page = await newSpecPage({
+    components: [Footer],
+    html: `<bm-footer>
+             <p>Hello, I'm in a footer!</p>
+           </bm-footer>`,
+  });
+
+  expect(page.root).toEqualHtml(`
+    <bm-footer>
+      <footer class="footer">
+        <p>Hello, I'm in a footer!</p>
+      </footer>
+    </bm-footer>
+  `);
+});