+++ packages/core/src/core/layout/section/test/section.spec.tsx
@@ -1,0 +1,17 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { Section } from '../section';
+
+it('Should render properly', async (): Promise<void> => {
+  const page = await newSpecPage({
+    components: [Section],
+    html: `<bm-section size="is-large">
+      <p>Hello, I'm in a section!</p>
+    </bm-section>`,
+  });
+
+  expect(page.root).toEqualHtml(`
+    <bm-section size="is-large" class="section is-large">
+      <p>Hello, I'm in a section!</p>
+    </bm-section>
+  `);
+});