+++ packages/react/src/components.ts
@@ -1,0 +1,50 @@
+/* eslint-disable */
+/* tslint:disable */
+/* auto-generated react proxies */
+import { createReactComponent } from './react-component-lib';
+
+import type { JSX } from '@bulmil/core';
+
+
+
+export const BmBadge = /*@__PURE__*/createReactComponent<JSX.BmBadge, HTMLBmBadgeElement>('bm-badge');
+export const BmBox = /*@__PURE__*/createReactComponent<JSX.BmBox, HTMLBmBoxElement>('bm-box');
+export const BmBreadcrumb = /*@__PURE__*/createReactComponent<JSX.BmBreadcrumb, HTMLBmBreadcrumbElement>('bm-breadcrumb');
+export const BmButton = /*@__PURE__*/createReactComponent<JSX.BmButton, HTMLBmButtonElement>('bm-button');
+export const BmButtons = /*@__PURE__*/createReactComponent<JSX.BmButtons, HTMLBmButtonsElement>('bm-buttons');
+export const BmCard = /*@__PURE__*/createReactComponent<JSX.BmCard, HTMLBmCardElement>('bm-card');
+export const BmCardContent = /*@__PURE__*/createReactComponent<JSX.BmCardContent, HTMLBmCardContentElement>('bm-card-content');
+export const BmCardFooter = /*@__PURE__*/createReactComponent<JSX.BmCardFooter, HTMLBmCardFooterElement>('bm-card-footer');
+export const BmCardHeader = /*@__PURE__*/createReactComponent<JSX.BmCardHeader, HTMLBmCardHeaderElement>('bm-card-header');
+export const BmCardImage = /*@__PURE__*/createReactComponent<JSX.BmCardImage, HTMLBmCardImageElement>('bm-card-image');
+export const BmCheckbox = /*@__PURE__*/createReactComponent<JSX.BmCheckbox, HTMLBmCheckboxElement>('bm-checkbox');
+export const BmColumn = /*@__PURE__*/createReactComponent<JSX.BmColumn, HTMLBmColumnElement>('bm-column');
+export const BmColumns = /*@__PURE__*/createReactComponent<JSX.BmColumns, HTMLBmColumnsElement>('bm-columns');
+export const BmContainer = /*@__PURE__*/createReactComponent<JSX.BmContainer, HTMLBmContainerElement>('bm-container');
+export const BmContent = /*@__PURE__*/createReactComponent<JSX.BmContent, HTMLBmContentElement>('bm-content');
+export const BmDivider = /*@__PURE__*/createReactComponent<JSX.BmDivider, HTMLBmDividerElement>('bm-divider');
+export const BmDropdown = /*@__PURE__*/createReactComponent<JSX.BmDropdown, HTMLBmDropdownElement>('bm-dropdown');
+export const BmField = /*@__PURE__*/createReactComponent<JSX.BmField, HTMLBmFieldElement>('bm-field');
+export const BmFile = /*@__PURE__*/createReactComponent<JSX.BmFile, HTMLBmFileElement>('bm-file');
+export const BmFooter = /*@__PURE__*/createReactComponent<JSX.BmFooter, HTMLBmFooterElement>('bm-footer');
+export const BmIcon = /*@__PURE__*/createReactComponent<JSX.BmIcon, HTMLBmIconElement>('bm-icon');
+export const BmImage = /*@__PURE__*/createReactComponent<JSX.BmImage, HTMLBmImageElement>('bm-image');
+export const BmInput = /*@__PURE__*/createReactComponent<JSX.BmInput, HTMLBmInputElement>('bm-input');
+export const BmMenu = /*@__PURE__*/createReactComponent<JSX.BmMenu, HTMLBmMenuElement>('bm-menu');
+export const BmMessage = /*@__PURE__*/createReactComponent<JSX.BmMessage, HTMLBmMessageElement>('bm-message');
+export const BmModal = /*@__PURE__*/createReactComponent<JSX.BmModal, HTMLBmModalElement>('bm-modal');
+export const BmNavbar = /*@__PURE__*/createReactComponent<JSX.BmNavbar, HTMLBmNavbarElement>('bm-navbar');
+export const BmNotification = /*@__PURE__*/createReactComponent<JSX.BmNotification, HTMLBmNotificationElement>('bm-notification');
+export const BmPagination = /*@__PURE__*/createReactComponent<JSX.BmPagination, HTMLBmPaginationElement>('bm-pagination');
+export const BmPanel = /*@__PURE__*/createReactComponent<JSX.BmPanel, HTMLBmPanelElement>('bm-panel');
+export const BmProgress = /*@__PURE__*/createReactComponent<JSX.BmProgress, HTMLBmProgressElement>('bm-progress');
+export const BmRadio = /*@__PURE__*/createReactComponent<JSX.BmRadio, HTMLBmRadioElement>('bm-radio');
+export const BmSection = /*@__PURE__*/createReactComponent<JSX.BmSection, HTMLBmSectionElement>('bm-section');
+export const BmSelect = /*@__PURE__*/createReactComponent<JSX.BmSelect, HTMLBmSelectElement>('bm-select');
+export const BmSlider = /*@__PURE__*/createReactComponent<JSX.BmSlider, HTMLBmSliderElement>('bm-slider');
+export const BmSwitch = /*@__PURE__*/createReactComponent<JSX.BmSwitch, HTMLBmSwitchElement>('bm-switch');
+export const BmTable = /*@__PURE__*/createReactComponent<JSX.BmTable, HTMLBmTableElement>('bm-table');
+export const BmTabs = /*@__PURE__*/createReactComponent<JSX.BmTabs, HTMLBmTabsElement>('bm-tabs');
+export const BmTag = /*@__PURE__*/createReactComponent<JSX.BmTag, HTMLBmTagElement>('bm-tag');
+export const BmTags = /*@__PURE__*/createReactComponent<JSX.BmTags, HTMLBmTagsElement>('bm-tags');
+export const BmTextarea = /*@__PURE__*/createReactComponent<JSX.BmTextarea, HTMLBmTextareaElement>('bm-textarea');