+++ packages/react/src/index.ts
@@ -1,0 +1,7 @@
+import { defineCustomElements } from '@bulmil/core/dist/loader';
+
+export * from './components';
+
+if (typeof window !== 'undefined') {
+  defineCustomElements(window);
+}