+++ CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/Gomah/bulmil/compare/v0.9.0...v0.10.0) (2020-09-01)
+
+
+### Bug Fixes
+
+* **react:** defineCustomElements in plugin ([1645b70](https://github.com/Gomah/bulmil/commit/1645b70a43d82d35950e9de68b32d03c228d6847))
+* **svelte:** defineCustomElements in plugin ([23cc609](https://github.com/Gomah/bulmil/commit/23cc6094fba38d5e883c9b5f4e664f0cb2760b5e))
+* **vue:** defineCustomElements in plugin ([f53c992](https://github.com/Gomah/bulmil/commit/f53c9924cfc1dba1718cc1329378d7e73b6206e9))
+
+
+### Features
+
+* **core:** new stencil config ([baaf75d](https://github.com/Gomah/bulmil/commit/baaf75db549425da50b386bd890b40a9b0b32c9b))
+
+
+
+
+
+# [0.9.0](https://github.com/Gomah/bulmil/compare/v0.8.0...v0.9.0) (2020-09-01)
+
+
+### Bug Fixes
+
+* **core:** use default for primary variable ([ba047b6](https://github.com/Gomah/bulmil/commit/ba047b602f689b915ac4401b2670bf53820e7644))
+* **react:** typo in package name ([105f64d](https://github.com/Gomah/bulmil/commit/105f64d6a200d38e671a7cae93903282577df2cd))
+
+
+### Features
+
+* **core:** bump stencil to 2.0.0.1 ([880f623](https://github.com/Gomah/bulmil/commit/880f623c56ffe771df7b226cb1d297badc6a29ad))
+* **core:** remove es5 build ([d3793c0](https://github.com/Gomah/bulmil/commit/d3793c06159d6ead45c8e7d17782ba61b5ec6cd1))
+
+
+
+
+
+# [0.8.0](https://github.com/Gomah/bulmil/compare/v0.7.6-alpha.0...v0.8.0) (2020-08-31)
+
+
+### Features
+
+* **stencil:** upgrade to stencil v2 ([76dd48b](https://github.com/Gomah/bulmil/commit/76dd48b0b76cb216e54f7401f18d7dcc19b7213b))
+
+
+
+
+
 ## [0.7.6-alpha.0](https://github.com/Gomah/bulmil/compare/v0.7.5-alpha.0...v0.7.6-alpha.0) (2020-08-31)